Move error injection structures to compilation unit
Required for #86
